% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_observed_residuals}
\alias{append_observed_residuals}
\title{Observed Residual Correlations}
\usage{
append_observed_residuals(data, na.rm = T)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{na.rm}{logical (defaults to TRUE)}
}
\value{
A numeric matrix of correlations among variable residuals.
}
\description{
Compute observed residual correlation (ORC) matrix among observed residuals
for variables supplied data.
}
\examples{
{
# Use the SCWB data example
data(SCWB)
append_observed_residuals(SCWB[,1:20])
}
}
