% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cor.rowMeans}
\alias{cor.rowMeans}
\title{Compute rowMeans of a correlation matrix}
\usage{
cor.rowMeans(x, na.rm = TRUE)
}
\arguments{
\item{x}{A correlations}

\item{na.rm}{logical (defaults to TRUE)}
}
\value{
A numeric vector of correlation means.
}
\description{
A wrapper for computing the mean of the rows of a correlation matrix.
}
\examples{
  # Use the SCWB data example
  data(SCWB)
  mycor <- cor(SCWB)
  cor.rowMeans(mycor)
}
