% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariances.R
\name{cov.rec}
\alias{cov.rec}
\title{Relative Excess Covariances}
\usage{
cov.rec(data, na.rm = TRUE)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{na.rm}{logical (defaults to TRUE)}
}
\value{
A numeric matrix of correlations among variable residuals.
}
\description{
Compute relative excess covariances (REC) matrix among variables in supplied
data.
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
{
  # Use the SCWB data example
  data(SCWB)
  cov.rec(SCWB[, 1:20])
}
}
