% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCWB}
\alias{SCWB}
\title{Subjective Community Well-Being Dataset (25 Variables)}
\format{
## `SCWB` A data frame with 501 observations of 25 variables.
\describe{
  \item{\code{cwbcloserel}}{}
  \item{\code{cwbrespect}}{}
  \item{\code{cwbtrust}}{}
  \item{\code{cwbmutuality}}{}
  \item{\code{cwbbeneficence}}{}
  \item{\code{cwbintegrity}}{}
  \item{\code{cwbcompetence}}{}
  \item{\code{cwbvision}}{}
  \item{\code{cwbrelgrowth}}{}
  \item{\code{cwbfairness}}{}
  \item{\code{cwbsustenance}}{}
  \item{\code{cwbachieve}}{}
  \item{\code{cwbsatisf}}{}
  \item{\code{cwbvalue}}{}
  \item{\code{cwbbelonging}}{}
  \item{\code{cwbwelcome}}{}
  \item{\code{cwbpurpose}}{}
  \item{\code{cwbcontrib}}{}
  \item{\code{cwbinterconn}}{}
  \item{\code{cwbsynergy}}{}
  \item{\code{total_goodrel}}{Mean of items cwbcloserel, cwbrespect, cwbtrust, cwbmutuality}
  \item{\code{total_prolead}}{Mean of items cwbbeneficence, cwbintegrity, cwbcompetence, cwbvision}
  \item{\code{total_healthyprac}}{Mean of items cwbrelgrowth, cwbfairness, cwbsustenance, cwbachieve}
  \item{\code{total_satcomm}}{Mean of items cwbsatisf, cwbvalue, cwbbelonging, cwbwelcome}
  \item{\code{total_strongmiss}}{Mean of items cwbpurpose, cwbcontrib, cwbintercon, cwbsynergy}
}
}
\source{
Data used in Padgett et al. (in press).
}
\usage{
data(SCWB)
}
\description{
These Subjective Community Well-Being (SCWB) data are a random subset from
the full dataset used in the paper by Padgett et al. (in press). The SCWB
dataset contains 501 responses to the 20 SCWB items plus 5 pre-computed total
scores on the five theorized domains of Good Relationships, Proficient
Leadership, Healthy Practices, Satisfying Community, and Strong Mission.
}
\examples{
# SCWB example used in preprint
data(SCWB)
}
\references{
Padgett, R.N., ...
}
\keyword{datasets}
