% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_observed_residuals}
\alias{append_observed_residuals}
\title{Append Observed Correlations to Data}
\usage{
append_observed_residuals(data, na.rm = TRUE)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{na.rm}{logical (defaults to TRUE)}
}
\value{
A numeric matrix of correlations among variable residuals.
}
\description{
Compute observed residuals for each variables in data and appends these newly
created variables to the original dataset.
}
\examples{
  # Use the SCWB data example
  data(SCWB)
  append_observed_residuals(SCWB[, 1:20])
}
