% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor.orc}
\alias{cor.orc}
\title{Observed Residual Correlations}
\usage{
cor.orc(data, na.rm = TRUE)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{na.rm}{logical (defaults to TRUE)}
}
\value{
A numeric matrix of correlations among variable residuals.
}
\description{
Compute observed residual correlation (ORC) matrix among observed residuals
for variables supplied data.
}
\examples{
{
  # Use the SCWB data example
  data(SCWB)
  cor.orc(SCWB[, 1:20])
}
}
