% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{cor.rec}
\alias{cor.rec}
\title{Relative Excess Correlations}
\usage{
cor.rec(data, na.rm = TRUE)
}
\arguments{
\item{data}{A data.frame or tibble}

\item{na.rm}{logical (defaults to TRUE)}
}
\value{
A numeric matrix of correlations among variable residuals.
}
\description{
Compute relative excess correlation (REC) matrix among variables in supplied
data.
}
\examples{
{
  # Use the SCWB data example
  data(SCWB)
  cor.rec(SCWB[, 1:20])
}
}
