% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmetrics-package.R
\docType{package}
\name{recmetrics-package}
\alias{recmetrics}
\alias{recmetrics-package}
\title{recmetrics: Psychometric Evaluation Using Relative Excess Correlations}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: R. Noah Padgett \email{npadgett@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-9114-3896}{ORCID}) [copyright holder]

}
\keyword{internal}
