% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recmetrics-package.R
\docType{package}
\name{recmetrics-package}
\alias{recmetrics}
\alias{recmetrics-package}
\title{recmetrics: Psychometric Evaluation Using Relative Excess Correlations}
\description{
Modern results of psychometric theory are implementedto provide users with a way of evaluating the internal structure of a set of items guided by theory. These methods are discussed in detail in VanderWeele and Padgett (2024) \doi{10.31234/osf.io/rnbk5}. The relative excess correlation matrices will, generally,have numerous negative entries even if all of the raw correlations between each pair of indicators are positive. The positive deviations of the relative excess correlation matrix entries help identify clusters of indicators that are more strongly related to one another, providing insights somewhat analogous to factor analysis, but without the need for rotations or decisions concerning the number of factors. A goal similar to exploratory/confirmatory factor analysis, but 'recmetrics' uses novel methods that do not rely on assumptions of latent variables or latent variable structures.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://noah-padgett.github.io/recmetrics/}
}

}
\author{
\strong{Maintainer}: R. Noah Padgett \email{npadgett@hsph.harvard.edu} (\href{https://orcid.org/0000-0002-9114-3896}{ORCID}) [copyright holder]

}
\keyword{internal}
